+++ Solution/RiskOfChaos/Utilities/VectorUtils.cs
@@ -1,38 +1,17 @@
-﻿using System;
-using UnityEngine;
+﻿using UnityEngine;
 
 namespace RiskOfChaos.Utilities
 {
     public static class VectorUtils
     {
-        static Vector3 cumulateVectorFunction(Vector3[] values, Func<Vector3, Vector3, Vector3> func)
+        public static Vector3 Spread(Vector3 direction, float maxAngle, Xoroshiro128Plus rng)
         {
-            if (values is null)
-                throw new ArgumentNullException(nameof(values));
-
-            if (values.Length == 0)
-                throw new ArgumentException($"{nameof(values)} must have at least one element", nameof(values));
-
-            if (values.Length == 1)
-                return values[0];
-
-            Vector3 result = func(values[0], values[1]);
-            for (int i = 2; i < values.Length; i++)
-            {
-                result = func(result, values[i]);
-            }
-
-            return result;
+            return Spread(direction, 0f, maxAngle, rng);
         }
 
-        public static Vector3 Min(params Vector3[] values)
+        public static Vector3 Spread(Vector3 direction, float minAngle, float maxAngle, Xoroshiro128Plus rng)
         {
-            return cumulateVectorFunction(values, Vector3.Min);
-        }
-
-        public static Vector3 Max(params Vector3[] values)
-        {
-            return cumulateVectorFunction(values, Vector3.Max);
+            return QuaternionUtils.Spread(direction, minAngle, maxAngle, rng) * direction;
         }
     }
 }